refactor(native): move the evidence-counting sink into its own module
`EvidenceToolCounts`, `EvidenceCountingSink` and its two impls move to
`api_process_adapter/sinks.rs`. `EvidenceCountingSink::{new, counts}` and
`EvidenceToolCounts`'s two counters become `pub(super)` because
`run_generation` and `project_native_outcomes` reach through them; every body
is byte-identical to its pre-split text.

The `agent_runtime/infrastructure` subtree budget rises to 58,357, sized for
the whole split rather than for this one module, so each commit in the series
lands green. The comment on the entry accounts for all 285 lines: they are
import blocks, module declarations, module docs, and two rustfmt rewraps.
